+++ _AppSetup.ps1
@@ -1,0 +1,28 @@
+Write-Host ---------------------------------------------------------------------------------
+Write-Host -------------------------------GOODBYE DPI---------------------------------------
+Write-Host ---------------------------------------------------------------------------------
+C:\PortableApps\goodbyedpi\service_install_any_dnsredir.cmd
+Write-Host .
+Write-Host .
+Start-Sleep -Seconds 2
+Write-Host .
+Write-Host .
+Write-Host ---------------------------------------------------------------------------------
+Write-Host -------------------------------SETUP SPOTIFY-------------------------------------
+Write-Host ---------------------------------------------------------------------------------
+C:\PortableApps\Spotify\start.bat
+Write-Host .
+Write-Host .
+Start-Sleep -Seconds 10
+Write-Host .
+Write-Host .
+<# Write-Host ---------------------------------------------------------------------------------
+Write-Host -----------------------------------FIREFOX TO DEFAULT----------------------------
+Write-Host ---------------------------------------------------------------------------------
+K:\FirefoxPortable\FirefoxPortable2DefaultPrograms.bat
+Write-Host .
+Write-Host .
+Start-Sleep -Seconds 10
+Write-Host .
+Write-Host . #> 
+Exit