+++ _DebloatScript.ps1
@@ -1,0 +1,42 @@
+Get-AppxPackage -allusers Disney.37853FC22B2CE | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.549981C3F5F10 | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.BingNews | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.BingWeather | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.GetHelp | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.Getstarted | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.Microsoft3DViewer | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.MicrosoftOfficeHub | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.MicrosoftSolitaireCollection | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.MicrosoftStickyNotes | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.MixedReality.Portal | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.MSPaint | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.Office.OneNote | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.OneDriveSync | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.People | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.PowerAutomateDesktop | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.SkypeApp | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.Todos | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.Wallet | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.WindowsAlarms | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.WindowsCamera | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.WindowsFeedbackHub | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.WindowsMaps | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.WindowsSoundRecorder | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.YourPhone | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.ZuneMusic | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.ZuneVideo | Remove-AppxPackage
+Get-AppxPackage -allusers MicrosoftTeams | Remove-AppxPackage
+Get-AppxPackage -allusers MicrosoftWindows.Client.WebExperience | Remove-AppxPackage
+Get-AppxPackage -allusers SpotifyAB.SpotifyMusic | Remove-AppxPackage
+Get-AppxPackage -allusers microsoft.windowscommunicationsapps | Remove-AppxPackage
+C:\Windows\SysWOW64\OneDriveSetup.exe -uninstall
+$winver = systeminfo | findstr /B /C:"OS Name"
+if ($winver -match 'Windows 10' -eq $true) {
+    C:\%PROGRAMFILES(X86)%\Microsoft\Edge\Application\9*\Installer\setup.exe --uninstall --force-uninstall --system-level # Windows 10 Microsoft Edge
+}
+if ($winver -match 'Windows 11' -eq $true) {
+    C:\%PROGRAMFILES(X86)%\Microsoft\Edge\Application\10*\Installer\setup.exe --uninstall --force-uninstall --system-level # Windows 11 Microsoft Edge
+}
+Get-WindowsPackage -Online | Where PackageName -like *Hello-Face* | Remove-WindowsPackage -Online -NoRestart
+Get-WindowsPackage -Online | Where PackageName -like *QuickAssist* | Remove-WindowsPackage -Online -NoRestart
+Exit