+++ Changelog.md
@@ -3,3 +3,10 @@
 1. Uninstall Microsoft Edge in Windows 11 
 2. Remove disabling Windows Defender in bat and reg files. Try to find better way
 3. Remove disable auto update. This block settings in Windows Update 
+
+# v0.7
+## Changes 
+1. Move to PowerShell
+2. Add ***_DebloatScript.ps1*** to uninstall blotware
+3. Add ***_AppSetup.ps1*** to setup my apps
+4. Move enviroment variable to script for better customization